Workaround for PHP 7 Compatibility
Archon's object composition technique makes heavy use of a PHP
feature allowing method calls from an incompatible context. However,
this ability was deprecated in PHP 5.6 and removed from PHP 7.0 [1].

Ultimately, it seems like moving over to Traits may be a logical path
forward.  This patch instead attempts to allow current Archon to run
on PHP 7+ with the fewest changes possible.

Rather than simply attempt the method call in the wrong context, we
instead create a closure in the desired class, then insert our current
context into that closure using bindTo.

There may, of course, be a simpler way to do this.  In particular, it
seems messy to need to create an object only to get the closure.  I
do not claim to be an expert in PHP internals, but this functions, and
is decently boiled down.

This new technique should be supported for PHP >= 5.4.0, but this is
untested.  One questionable area would be the use of the getClosure
method call on a parenthesized "new" object statement.  I believe it is
supported syntax on at least 5.6, and would be very easy to work around
in any case.  That said, it may be better to retain the old technique
and branch within the code to suit the running version.

[1] http://php.net/manual/en/migration56.deprecated.php#migration56.deprecated.incompatible-context

Signed-off-by: Dan Wells <dbw2@calvin.edu>